Restructure CoreExt
