+++ lib/atomic_time/core_ext/active_record.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
-require 'active_record'
 require_relative 'active_record/base'
-require_relative 'active_record/connection_adapters'
 require_relative 'active_record/type'
