--- lib/atomic_time/core_ext/active_record/connection_adapters.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require 'active_record'
-require_relative 'connection_adapters/column_methods'
