--- lib/atomic_time/core_ext/active_record/connection_adapters/column_methods.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-require 'active_record'
-
-module ActiveRecord
-  module ConnectionAdapters
-    module ColumnMethods
-      def atomic_time(*args, **options)
-        args.each { |name| column(name, :integer, options) }
-      end
-    end
-  end
-end
