+++ lib/atomic_time/core_ext/date_and_time/conversions.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class AtomicTime
+  module CoreExt
+    module DateAndTime
+      module Conversions
+        def to_atomic_time
+          AtomicTime.new hour, min, sec
+        end
+      end
+    end
+  end
+end
