+++ lib/atomic_time/core_ext/date_time.rb
@@ -1,9 +1,3 @@
 # frozen_string_literal: true
 
-require 'date'
-
-class DateTime
-  def to_atomic_time
-    AtomicTime.new hour, min, sec
-  end
-end
+require_relative 'date_time/conversions'
