+++ lib/atomic_time/core_ext/string.rb
@@ -1,7 +1,3 @@
 # frozen_string_literal: true
 
-class String
-  def to_atomic_time
-    AtomicTime.parse self
-  end
-end
+require_relative 'string/conversions'
