+++ lib/atomic_time/core_ext/time.rb
@@ -1,7 +1,3 @@
 # frozen_string_literal: true
 
-class Time
-  def to_atomic_time
-    AtomicTime.new hour, min, sec
-  end
-end
+require_relative 'time/conversions'
