+++ lib/atomic_time/core_ext/time/conversions.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'atomic_time/core_ext/date_and_time/conversions'
+
+Time.include AtomicTime::CoreExt::DateAndTime::Conversions
