+++ lib/umts_time.rb
@@ -2,4 +2,3 @@
 
 require 'atomic_time'
 require_relative 'umts_time/core_ext'
-require_relative 'umts_time/formats'
