+++ lib/umts_time/core_ext.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative 'core_ext/active_support'
 require_relative 'core_ext/atomic_time'
 require_relative 'core_ext/date'
 require_relative 'core_ext/date_time'
+require_relative 'core_ext/duration'
 require_relative 'core_ext/range'
 require_relative 'core_ext/time'
