--- lib/umts_time/core_ext/active_support.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require 'active_support'
-require_relative 'active_support/duration'
