--- lib/umts_time/core_ext/active_support/duration.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require 'active_support/duration'
-
-module ActiveSupport
-  class Duration
-    def convert_to(unit)
-      to_f / PARTS_IN_SECONDS[unit]
-    end
-  end
-end
