+++ lib/umts_time/core_ext/atomic_time/am_parser.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'atomic_time'
+require 'umts_time/core_ext/date_and_time/am_parser'
+
+AtomicTime.extend UMTSTime::CoreExt::DateAndTime::AmParser
