+++ lib/umts_time/core_ext/atomic_time/formats.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'atomic_time'
+require 'umts_time/formats'
+
+AtomicTime::DATE_FORMATS = UMTSTime::Formats::DATE_TIME
