+++ lib/umts_time/core_ext/date.rb
@@ -1,10 +1,4 @@
 # frozen_string_literal: true
 
-require 'date'
-require 'umts_time/american_parser'
-require 'umts_time/formats'
-
-class Date
-  include UMTSTime::Formats
-  include UMTSTime::AmericanParser
-end
+require_relative 'date/am_parser'
+require_relative 'date/formats'
