+++ lib/umts_time/core_ext/date/formats.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'active_support/core_ext/date/conversions'
+require 'date'
+require 'umts_time/formats'
+
+Date.send :remove_const, :DATE_FORMATS
+Date::DATE_FORMATS = UMTSTime::Formats::DATE
