+++ lib/umts_time/core_ext/date_time.rb
@@ -1,10 +1,4 @@
 # frozen_string_literal: true
 
-require 'date'
-require 'umts_time/american_parser'
-require 'umts_time/formats'
-
-class DateTime
-  include UMTSTime::AmericanParser
-  include UMTSTime::Formats
-end
+require_relative 'date_time/am_parser'
+require_relative 'date_time/formats'
