+++ lib/umts_time/core_ext/date_time/formats.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'date'
+require 'umts_time/formats'
+
+DateTime::DATE_FORMATS = UMTSTime::Formats::DATE_TIME
