+++ lib/umts_time/core_ext/duration.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative 'duration/conversions'
