+++ lib/umts_time/core_ext/time.rb
@@ -1,9 +1,4 @@
 # frozen_string_literal: true
 
-require 'umts_time/american_parser'
-require 'umts_time/formats'
-
-class Time
-  include UMTSTime::Formats
-  include UMTSTime::AmericanParser
-end
+require_relative 'time/am_parser'
+require_relative 'time/formats'
