+++ lib/umts_time/core_ext/time/am_parser.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'umts_time/core_ext/date_and_time/am_parser'
+
+Time.extend UMTSTime::CoreExt::DateAndTime::AmParser
